Tests/LibWeb/Text/expected/html-form-controls-collection.txt

      HTMLFormControlsCollection
RadioNodeList
2
button
text
null
HTMLInputElement
text
